Add $post_id to AudioTheme record links filter.
In some cases, the passed post ID is needed when filtering the links.

For example, if the records links are shown in a widget (e.g. custom Record widget output [Huesos]), the filtered url results may not be correct without the use of the passed $post object ID.

I tested this using our AudioTheme Cart Connector plugin and the Huesos theme. To get the record links to work, an update to that plugin needs applied, too (commit pushed). This can be tested by using that plugin, the Huesos theme, and adding a Record Widget to the home page widget area. The record links will display, however, the EDD url will not work without these updates.